move internal code into framework namespace
