+++ GiftTasteHelper/GiftTasteHelper.csproj
@@ -37,20 +37,20 @@
     <Reference Include="System" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Calendar.cs" />
-    <Compile Include="GiftHelperType.cs" />
+    <Compile Include="Framework\Calendar.cs" />
+    <Compile Include="Framework\GiftHelperType.cs" />
     <Compile Include="GiftTasteHelper.cs" />
-    <Compile Include="CalendarGiftHelper.cs" />
-    <Compile Include="GiftHelper.cs" />
-    <Compile Include="IGiftHelper.cs" />
-    <Compile Include="ItemData.cs" />
-    <Compile Include="ModConfig.cs" />
-    <Compile Include="NPCGiftInfo.cs" />
+    <Compile Include="Framework\CalendarGiftHelper.cs" />
+    <Compile Include="Framework\GiftHelper.cs" />
+    <Compile Include="Framework\IGiftHelper.cs" />
+    <Compile Include="Framework\ItemData.cs" />
+    <Compile Include="Framework\ModConfig.cs" />
+    <Compile Include="Framework\NPCGiftInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SocialPage.cs" />
-    <Compile Include="SocialPageGiftHelper.cs" />
-    <Compile Include="SVector2.cs" />
-    <Compile Include="Utils.cs" />
+    <Compile Include="Framework\SocialPage.cs" />
+    <Compile Include="Framework\SocialPageGiftHelper.cs" />
+    <Compile Include="Framework\SVector2.cs" />
+    <Compile Include="Framework\Utils.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="manifest.json">
